Fix the grid module: subgrid flags, database handling and CSV output
From the adversarial review of the grid chunk.

The subgrid a plot belongs to was stored as 1 << distance, and the
distances go up to 100 while an int has 32 bits: the grid of 50 landed
on the bit of the grid of 18 and the grid of 100 on the bit of the grid
of 4. AbstractStore.getGridFlags now uses one bit per position of the
distance list, which always fits, and both stores call it. The queries
no longer read gridflags at all -- they test the row and the column
directly, which is what belonging to a subgrid means and which also
gives the right answer for rows written by the old encoding.

JDBCStore: a failed connection was logged and null returned, so every
caller died on a NullPointerException; a failed insert batch dropped up
to 50000 plots and the run carried on; getAllStatement cached its SQL in
the field of getSelectStatement, so whichever query was asked for second
answered with the other one's SQL; closeStore opened a connection and
prepared the statements again only to close them; the table script was
read as a File, which throws as soon as the module runs from its jar;
and the coordinates were rounded through a float, losing about a metre.
Results are now streamed rather than loaded whole.

CSVStore: when zipping, a plain FileWriter was opened on the same file
and left open, truncating what the zip stream then wrote; the rows were
two columns wider than the header; a failure opening the files was
logged and left the writers null; and closeStore ignored the writer's
own error flag.

Also: HibernateStore flushed on every plot of row 0 and never again,
keeping the whole run in memory; the row and griddistance columns were
smallint, too small for a griddistance of 100000 or 200000; Hibernate
was configured to drop and recreate the plot table on every start, under
a property name it no longer reads; Plot compared the coordinates and
the flags although only the three key fields identify a plot; and
QuerySigrid left cursors open and could skip closing the database.

New tables carry the new gridflags encoding, which is not the old one;
existing rows are unaffected since nothing reads that column any more.

+++ collect-earth/collect-earth-grid/src/main/resources/createTableSqlite.sql
@@ -4,8 +4,8 @@
 
 CREATE TABLE IF NOT EXISTS plot
 (
-  row smallint NOT NULL,
-  griddistance smallint NOT NULL,
+  row integer NOT NULL,
+  griddistance integer NOT NULL,
   col integer NOT NULL,
   gridflags integer,
   xcoordinate integer,
